Add files via upload
In this code, I cleared all outputs because GitHub cannot import results when the code contains metadata. Widgets.
Hence, I uploaded the code without results and ran it on Google Colab Pro with an A100 GPU. Users should first add credit to their Google Colab accounts to upgrade to the Pro version, then use the A100 GPU. Another thing that comes to mind is that you need to get your Kaggle token for Food101. Dataset, and upload it into the code. However, I completed all these steps, and for your convenience, I uploaded the code, its results, and the PDF report of the code execution in a ZIP file, separately.