Add files via upload
The executed code with its results was uploaded. You can download this Zip file, extract the “.ipynb" file, and upload it again to your Google Colab or Jupyter notebook to check the results.
Just so you know, if you want to run this project, you will need to use Google Colab Pro with an A100 GPU, so you need to add credit to your Google Colab account.

Enjoy it!